Cool so that's bio & English out of the way
There are two types of slackers: those who admit it, and those who claim they'll do the work and, even if they do, give extremely low results. That's what I have to deal with in AP English. No names of course, but he's made so may promises and delivered only like 10% of what I've asked. For reference, I've done probally 90% of the work so far. This last weekend I asked him to come up with only two counters and two Q&A answers, and you know what he does? He scrapes together enough in the morning before school and at lunch so that when I confronted him he was like "Oh no, I did do it".

Like bro, please.
I really don't want to have to carry us. But hey, it's what I get for not making friends.

Anyways, peace out~

+++ net/keinr/util/config/Properties.java
@@ -57,7 +57,7 @@ public class Properties {
      * @throws NumberFormatException if the property isn't a long
      * @return property long value
      */
-    public int getLong(String tag, String name) throws PropertyNFException {
+    public long getLong(String tag, String name) throws PropertyNFException {
         return Long.parseLong(getString(tag, name));
     }
